# .docs/.openapi/api.yaml
openapi: 3.0.3
info:
  contact:
    email: andreas.rauber@tuwien.ac.at
    name: Prof. Andreas Rauber
  description: >
    The merged REST API of DBRepo for users, developers and data stewards to be
    accessed publicly. Have a look at

    the [source
    code](https://gitlab.phaidra.org/fair-data-austria-db-repository/fda-services)
    for non-public endpoints

    that are used between the services themselves.
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
  title: DBRepo REST API
  version: 1.6.1
servers:
  - description: Test Instance
    url: 'https://test.dbrepo.tuwien.ac.at'
  - description: Local Instance
    url: 'http://localhost'
...
paths:
...
  '/api/database/{databaseId}/subset/{subsetId}':
    get:
      tags:
        - subset-endpoint
      summary: Find subset
      description: >-
        Finds a subset in the data database.  When the database schema is marked
        as hidden, the user needs to be authorized, have at least read-access to
        the database.  Requests with HTTP header `Accept=application/json`
        return the metadata, requests with HTTP header `Accept=text/csv` return
        the data as downloadable file.
      operationId: findById
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: subsetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: Accept
          in: header
          required: true
          schema:
            type: string
        - name: timestamp
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Found subset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryDto'
            text/csv: {}
        '400':
          description: Malformed select query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Not allowed to find subset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: >-
            Failed to find database in metadata database or query in query store
            of the data database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '406':
          description: Failed to find acceptable representation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: Failed to communicate with database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - basicAuth: []
        - bearerAuth: []
...
  '/api/database/{databaseId}/access/{userId}':
    get:
      tags:
        - access-endpoint
      summary: Find/Check access
      description: >-
        Finds or checks access of a user with given id to a database with given
        id. Requests with HTTP method **GET** return the access object, requests
        with HTTP method **HEAD** only the status. When the user has at least
        *READ* access, the status 200 is returned, 403 otherwise. Requires role
        `check-database-access` or `check-foreign-database-access`.
      operationId: find
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Found database access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseAccessDto'
        '403':
          description: No access to this database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Database not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
    put:
      tags:
        - access-endpoint
      summary: Modify access
      description: >-
        Modifies access of a user with given id to database with given id.
        Requires role `update-database-access`.
      operationId: update_5
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessDto'
        required: true
      responses:
        '202':
          description: Modified access
        '400':
          description: Modify access query or database connection is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: >-
            Modify access not permitted when no access is granted in the first
            place
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Database or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '502':
          description: >-
            Access could not be updated due to connection error in the data
            service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: Access could not be updated in the data service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
    post:
      tags:
        - access-endpoint
      summary: Give access
      description: >-
        Give a user with given id access to some database with given id.
        Requires role `create-database-access`.
      operationId: create_8
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessDto'
        required: true
      responses:
        '202':
          description: Granting access succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseAccessDto'
        '400':
          description: Granting access query or database connection is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Failed giving access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Database or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '502':
          description: Access could not be created due to connection error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: Access could not be created in the data service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
    delete:
      tags:
        - access-endpoint
      summary: Delete access
      description: >-
        Delete access of a user with id to a database with id. Requires role
        `delete-database-access`.
      operationId: revoke
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Deleted access
        '400':
          description: Modify access query or database connection is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Revoke of access not permitted as no access was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: 'User, database with access was not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '502':
          description: Access could not be created due to connection error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: Access could not be revoked in the data service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
    head:
      tags:
        - access-endpoint
      summary: Find/Check access
      description: >-
        Finds or checks access of a user with given id to a database with given
        id. Requests with HTTP method **GET** return the access object, requests
        with HTTP method **HEAD** only the status. When the user has at least
        *READ* access, the status 200 is returned, 403 otherwise. Requires role
        `check-database-access` or `check-foreign-database-access`.
      operationId: find_1
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Found database access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseAccessDto'
        '403':
          description: No access to this database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Database not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
...
  '/api/database/{databaseId}':
    get:
      tags:
        - database-endpoint
      summary: Find database
      description: Finds a database with id.
      operationId: findById_1
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Database found successfully
          headers:
            X-Username:
              description: The authentication username
              style: simple
            Access-Control-Expose-Headers:
              description: Expose custom headers
              style: simple
            X-Password:
              description: The authentication password
              style: simple
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseBriefDto'
        '403':
          description: Not allowed to view database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: 'Database, user or exchange could not be found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '502':
          description: Connection to the broker service could not be established
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: Failed to find queue information in broker service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
...
  '/api/container/{containerId}':
    get:
      tags:
        - container-endpoint
      summary: Find container
      description: Finds a container in the metadata database.
      operationId: findById_3
      parameters:
        - name: containerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Found container
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerDto'
        '404':
          description: Container image could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
    delete:
      tags:
        - container-endpoint
      summary: Delete container
      description: >-
        Deletes a container in the metadata database. Requires role
        `delete-container`.
      operationId: delete_6
      parameters:
        - name: containerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: Deleted container
        '403':
          description: 'Create container not permitted, need authority `delete-container`'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Container not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
...
components:
  securitySchemes:
    basicAuth:
      in: header
      scheme: basic
      type: http
    bearerAuth:
      bearerFormat: JWT
      in: header
      scheme: bearer
      type: http
  schemas:
...
    QueryDto:
      required:
        - database_id
        - execution
        - id
        - identifiers
        - is_persisted
        - owner
        - query
        - query_hash
        - query_normalized
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 4
        owner:
          $ref: '#/components/schemas/UserBriefDto'
        execution:
          type: string
          format: date-time
          example: '2021-03-12T15:26:21.000Z'
        query:
          type: string
          example: SELECT `id` FROM `air_quality`
        type:
          type: string
          example: query
          enum:
            - query
            - view
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/IdentifierBriefDto'
        database_id:
          type: integer
          format: int64
          example: 1
        query_normalized:
          type: string
          example: SELECT `id` FROM `air_quality`
        query_hash:
          type: string
          example: 17e682f060b5f8e47ea04c5c4855908b0a5ad612022260fe50e11ecb0cc0ab76
        is_persisted:
          type: boolean
          example: true
        result_hash:
          type: string
          example: 17e682f060b5f8e47ea04c5c4855908b0a5ad612022260fe50e11ecb0cc0ab76
        result_number:
          type: integer
          format: int64
          example: 1
...
    DatabaseBriefDto:
      required:
        - contact
        - id
        - identifiers
        - internal_name
        - is_public
        - is_schema_public
        - name
        - owner_id
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        name:
          type: string
          example: Air Quality
        description:
          type: string
          example: Air Quality
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/IdentifierBriefDto'
        contact:
          $ref: '#/components/schemas/UserBriefDto'
        internal_name:
          type: string
          example: air_quality
        is_public:
          type: boolean
          example: true
        is_schema_public:
          type: boolean
          example: true
        owner_id:
          type: string
          format: uuid
          example: 2f45ef7a-7f9b-4667-9156-152c87fe1ca5
        preview_image:
          type: string
    DatabaseAccessDto:
      required:
        - type
        - user
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserBriefDto'
        type:
          type: string
          example: read
          enum:
            - read
            - write_own
            - write_all
...
    CreateAccessDto:
      required:
        - type
      type: object
      properties:
        type:
          type: string
          example: read
          enum:
            - read
            - write_own
            - write_all
...